feat(production-pulse): permitir intervalo mínimo de poll de 0,5 s.
Migration V002 com NUMERIC no banco, validação API/MFE alinhada e testes de subsegundo no scheduler.

+++ production-pulse-api/production_pulse_app/application/services/device_poll_scheduler_service.py
@@ -98,7 +98,7 @@ class DevicePollSchedulerService:
                     self._in_flight.discard(device_id)
 
     def _schedule_next(self, device: dict[str, Any]) -> None:
-        next_at = compute_next_poll_at(int(device["poll_interval_seconds"]))
+        next_at = compute_next_poll_at(float(device["poll_interval_seconds"]))
         self._devices.update_next_poll_at(device["id"], next_poll_at=next_at)
